+++ package.json
@@ -1,10 +1,14 @@
 {
   "name": "todo-list",
+  "jest": {
+    "verbose": true,
+    "testEnvironment": "jsdom"
+},
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --coverage",
     "dev": "webpack --mode development",
     "start": "webpack serve --open"
   },
@@ -21,6 +25,7 @@
   "homepage": "https://github.com/looh16/todo-list#readme",
   "devDependencies": {
     "@babel/core": "^7.18.6",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
     "@babel/preset-env": "^7.18.6",
     "autoprefixer": "^10.4.7",
     "babel-eslint": "^10.1.0",
@@ -32,6 +37,9 @@
     "file-loader": "^6.2.0",
     "hint": "^7.1.0",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-localstorage-mock": "^2.4.21",
     "mini-css-extract-plugin": "^2.6.1",
     "postcss-loader": "^7.0.1",
     "sass": "^1.53.0",
