+++ src/update-status.js
@@ -14,20 +14,3 @@ export const updateToTrue = (todo) => {
   localStoragetodos[index].completed = true;
   localStorage.setItem('todos', JSON.stringify(localStoragetodos));
 };
-
-const btnClear = document.getElementById('deleteAllCompleted');
-// eslint-disable-next-line
-btnClear.addEventListener('click', (e) => {
-  let localStoragetodos = getAllTodos();
-  localStoragetodos = localStoragetodos.filter((todo) => todo.completed !== true);
-  localStorage.setItem('todos', JSON.stringify(localStoragetodos));
-  // eslint-disable-next-line
-  location.reload();
-});
-
-const btnRefresh = document.getElementById('refreshDom');
-// eslint-disable-next-line
-btnRefresh.addEventListener('click', (e) => {
-// eslint-disable-next-line
-  location.reload();
-});
